[omp] renamed *_torch.py files to *.py
