+++ coppafish/omp/spots.py
